requirements.txt
Django
djangorestframework
Markdown
django-environ
mysqlclient
Pillow
gunicorn